+++ tests/fixtures.test.ts
@@ -8,7 +8,6 @@ import { format, pluginPath } from './utils'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
-
 const execAsync = promisify(exec)
 
 let fixtures = [
@@ -70,7 +69,17 @@ let fixtures = [
   },
   {
     name: 'v4: configs and plugins',
-    dir: 'v4/configs',
+    dir: 'v4/css-loading-js',
+    ext: 'html',
+  },
+  {
+    name: 'custom npm package name: v3',
+    dir: 'custom-pkg-name-v3',
+    ext: 'html',
+  },
+  {
+    name: 'custom npm package name: v4',
+    dir: 'custom-pkg-name-v4',
     ext: 'html',
   },
 ]
@@ -117,8 +126,10 @@ describe('fixtures', () => {
 
     test.concurrent(name, async ({ expect }) => {
       let results = await execAsync(cmd)
-      let formatted = results.stdout
-      let expected = await fs.readFile(outputPath, 'utf-8')
+      let formatted = results.stdout.replace(/\r\n/g, '\n')
+      let expected = await fs
+        .readFile(outputPath, 'utf-8')
+        .then((c) => c.replace(/\r\n/g, '\n'))
 
       expect(formatted.trim()).toEqual(expected.trim())
     })